Refactor: changes some constants name (#1641)
* refactor: removing redeclaration constants

* refactor: changes constants name of event roles constants

* test: fix failing test for because events roles constants

* refactor: changes constants name of progress response message constants
