+++ constants/progresses.js
@@ -6,7 +6,7 @@ const MILLISECONDS_IN_DAY = 24 * 60 * 60 * 1000;
 const INTERNAL_SERVER_ERROR_MESSAGE =
   "The server has encountered an unexpected error. Please contact the administrator for more information.";
 
-const RESPONSE_MESSAGES = {
+const PROGRESSES_RESPONSE_MESSAGES = {
   PROGRESS_DOCUMENT_CREATED_SUCCEEDED,
   PROGRESS_DOCUMENT_RETRIEVAL_SUCCEEDED,
   PROGRESS_DOCUMENT_NOT_FOUND,
@@ -22,7 +22,7 @@ const PROGRESS_VALID_SORT_FIELDS = ["date", "-date"];
 const VALID_PROGRESS_TYPES = ["task", "user"];
 
 module.exports = {
-  RESPONSE_MESSAGES,
+  PROGRESSES_RESPONSE_MESSAGES,
   MILLISECONDS_IN_DAY,
   INTERNAL_SERVER_ERROR_MESSAGE,
   TYPE_MAP,
