+++ constants/users.js
@@ -17,8 +17,6 @@ const ALLOWED_FILTER_PARAMS = {
   ROLE: ["role"],
 };
 
-const DOCUMENT_WRITE_SIZE = 500;
-
 const FIRESTORE_IN_CLAUSE_SIZE = 30;
 
 const USERS_PATCH_HANDLER_ACTIONS = {
@@ -54,7 +52,6 @@ module.exports = {
   profileStatus,
   USER_STATUS,
   ALLOWED_FILTER_PARAMS,
-  DOCUMENT_WRITE_SIZE,
   FIRESTORE_IN_CLAUSE_SIZE,
   USERS_PATCH_HANDLER_ACTIONS,
   USERS_PATCH_HANDLER_ERROR_MESSAGES,
