+++ controllers/events.js
@@ -1,4 +1,4 @@
-const { GET_ALL_EVENTS_LIMIT_MIN, UNWANTED_PROPERTIES_FROM_100MS, ROLES } = require("../constants/events");
+const { GET_ALL_EVENTS_LIMIT_MIN, UNWANTED_PROPERTIES_FROM_100MS, EVENT_ROLES } = require("../constants/events");
 const { INTERNAL_SERVER_ERROR } = require("../constants/errorMessages");
 const { EventTokenService, EventAPIService } = require("../services");
 const eventQuery = require("../models/events");
@@ -98,7 +98,7 @@ const joinEvent = async (req, res) => {
   const payload = { userId, role };
 
   try {
-    if (role === ROLES.MAVEN) {
+    if (role === EVENT_ROLES.MAVEN) {
       const eventCodes = await eventQuery.getEventCodes({ id: roomId });
       const allEventCodesArray = eventCodes.map((eventCode) => {
         return eventCode.code;
@@ -119,14 +119,14 @@ const joinEvent = async (req, res) => {
       });
     }
 
-    if (role === ROLES.HOST || role === ROLES.MODERATOR) {
+    if (role === EVENT_ROLES.HOST || role === EVENT_ROLES.MODERATOR) {
       if (!req.userData) {
         return res.status(400).json({
           message: "Unauthorized, please login to perform this action!",
         });
       }
 
-      if (role === ROLES.HOST && req.userData.roles.super_user) {
+      if (role === EVENT_ROLES.HOST && req.userData.roles.super_user) {
         const token = tokenService.getAuthToken({ ...payload, roomId: roomId });
 
         return res.status(201).json({
@@ -135,7 +135,7 @@ const joinEvent = async (req, res) => {
         });
       }
 
-      if (role === ROLES.MODERATOR && req.userData.roles.member) {
+      if (role === EVENT_ROLES.MODERATOR && req.userData.roles.member) {
         const token = tokenService.getAuthToken({ ...payload, roomId: roomId });
 
         return res.status(201).json({
@@ -317,7 +317,7 @@ const generateEventCode = async (req, res) => {
   const { eventCode, role } = req.body;
   const eventCodeUuid = crypto.randomUUID({ disableEntropyCache: true });
 
-  if (role !== ROLES.MAVEN) {
+  if (role !== EVENT_ROLES.MAVEN) {
     return res.status(400).json({
       message: "Currently the room codes feature is only for mavens!",
     });
